+++ packages/app-bridge/src/AppBridgeBlock.ts
@@ -36,7 +36,9 @@ import type {
 } from './types';
 import { PrivacySettings } from './types/PrivacySettings';
 
-export type BlockApiMethod = ApiMethodNameValidator<Pick<ApiMethodRegistry, 'getCurrentUser'>>;
+export type BlockApiMethod = ApiMethodNameValidator<
+    Pick<ApiMethodRegistry, 'getAssetBulkDownloadToken' | 'getCurrentUser'>
+>;
 
 export type BlockCommand = CommandNameValidator<
     Pick<
@@ -46,6 +48,7 @@ export type BlockCommand = CommandNameValidator<
         | 'downloadAsset'
         | 'openAssetChooser'
         | 'openAssetViewer'
+        | 'openNewPublication'
         | 'openTemplateChooser'
     >
 >;
@@ -168,8 +171,6 @@ export interface AppBridgeBlock<
     closeTemplateChooser(): void;
 
     getCurrentLoggedUser(): Promise<User>;
-
-    getBulkDownloadToken(assetIds: number[], setIds?: number[]): Promise<string>;
 
     getBulkDownloadByToken(token: string): Promise<BulkDownload>;