+++ packages/guideline-blocks-settings/src/components/RichTextEditor/RichTextEditor.spec.ct.tsx
@@ -10,6 +10,7 @@ import { convertToRteValue } from '../../helpers';
 const RteHtmlSelector = '[data-test-id="rte-content-html"]';
 const RichTextSelector = '[data-test-id="rich-text-editor"]';
 const ButtonSelector = '[data-test-id="button"]';
+const CheckboxSelector = '[data-test-id="checkbox-icon-box"]';
 const ToolbarButtonSelector = '[data-testid="ToolbarButton"]';
 const InternalDocumentLinkSelector = '[data-test-id="internal-link-selector-document-link"]';
 const FloatingLinkModalSelector = '[data-test-id="floating-link-insert"]';
@@ -71,7 +72,33 @@ describe('RichTextEditor', () => {
         cy.get(InternalDocumentLinkSelector).click();
         cy.get(ButtonSelector).last().click();
         cy.get(FloatingLinkModalSelector).find(ButtonSelector).last().click();
-        cy.get(RichTextSelector).find('a[href="/r/document"]').should('exist');
+        const linkTag = cy.get(RichTextSelector).find('a[href="/r/document"]');
+        linkTag.should('exist');
+        linkTag.should('have.attr', 'target', '_self');
+    });
+
+    it('should be able to select internal link with target blank', () => {
+        (appBridge.getDocumentGroups as SinonStub) = cy.stub().returns([]);
+        (appBridge.getAllDocuments as SinonStub) = cy.stub().returns(Promise.resolve(apiDocuments));
+
+        mount(
+            <RichTextEditor
+                isEditing={true}
+                plugins={new PluginComposer().setPlugin([new LinkPlugin({ appBridge })])}
+                value={convertToRteValue('p', 'This is a link')}
+            />,
+        );
+        cy.get(RichTextSelector).click();
+        cy.get(RichTextSelector).type('{selectall}');
+        cy.get(ToolbarButtonSelector).click();
+        cy.get(ButtonSelector).first().click();
+        cy.get(InternalDocumentLinkSelector).click();
+        cy.get(ButtonSelector).last().click();
+        cy.get(CheckboxSelector).click();
+        cy.get(FloatingLinkModalSelector).find(ButtonSelector).last().click();
+        const linkTag = cy.get(RichTextSelector).find('a[href="/r/document"]');
+        linkTag.should('exist');
+        linkTag.should('have.attr', 'target', '_blank');
     });
 
     it('should prepend the URL with https:// if not exists', () => {