+++ packages/guideline-blocks-settings/src/components/RichTextEditor/plugins/ButtonPlugin/components/FloatingButton/EditButtonModal/EditModal.tsx
@@ -2,17 +2,17 @@
 
 import { IconPen16, IconTrashBin16 } from '@frontify/fondue';
 import { FloatingButton, useFloatingButtonUrlInput } from '..';
-import { BlockStyles } from '../../../..';
 
 export const EditModal = () => {
     const urlHtmlProps = useFloatingButtonUrlInput({});
 
     return (
-        <div data-test-id="floating-button-edit" className="tw-bg-white tw-rounded tw-shadow tw-p-4 tw-min-w-[400px]">
-            <span data-test-id="preview-button-flyout" className="tw-flex tw-justify-between">
-                <span className="tw-pointer-events-none" style={BlockStyles.p}>
-                    {urlHtmlProps.defaultValue}
-                </span>
+        <div
+            data-test-id="floating-button-edit"
+            className="tw-bg-white tw-text-text tw-rounded tw-shadow tw-p-4 tw-min-w-[400px]"
+        >
+            <span data-test-id="preview-button-flyout" className="tw-flex tw-justify-between tw-items-center">
+                <span className="tw-pointer-events-none">{urlHtmlProps.defaultValue}</span>
                 <span className="tw-flex tw-gap-2">
                     <span
                         role="button"